+++ test/integration/app-dir-export/test/dynamic-prod.test.ts
@@ -1,0 +1,1 @@
+require('./dynamic')
