+++ test/integration/app-dir-export/test/trailing-slash-dev.test.ts
@@ -1,0 +1,2 @@
+process.env.TEST_DEV = '1'
+require('./dynamic')
