+++ test/integration/app-dir-export/test/trailing-slash-start.test.ts
@@ -1,0 +1,1 @@
+require('./dynamic')
