+++ test/integration/app-dir-export/test/trailing-slash.ts
@@ -1,0 +1,13 @@
+import { runTests } from './utils'
+
+it.each([
+  { isDev: true, trailingSlash: false },
+  { isDev: true, trailingSlash: true },
+  { isDev: false, trailingSlash: false },
+  { isDev: false, trailingSlash: true },
+])(
+  "should work with isDev '$isDev' and trailingSlash '$trailingSlash'",
+  async ({ isDev, trailingSlash }) => {
+    await runTests({ isDev, trailingSlash })
+  }
+)
